ARM64 portability hardening: fix UB in interpolation, binning, and hot-path code
Audit-driven fixes for undefined behavior and x86/ARM divergences found on
aarch64 (verified with a 6-way parallel code audit + UBSan runs):

Bug fixes (observable):
- equi_width_bins: integer path hung forever (uninterruptible, surviving
  max_execution_time) when the bin step truncated to 0 - now throws
  InvalidInputException, like the double path
- InterpolateOperator (window FILL): casting an out-of-range double back to
  int64/uint64 is UB and diverges across platforms (x86 cvttsd2si yields the
  integer-indefinite value, ARM fcvtzs saturates) - clamp to the target range
  before casting
- quantile_cont over TIMESTAMP/DATE: same overflow via llround - compute in
  long double (exact for int64 on x86-64 and aarch64) and clamp, yielding the
  mathematically correct finite timestamp instead of +/-infinity
- lead/lag with INT64_MIN offset: negating/abs of INT64_MIN is UB - decline
  the streaming path for it (falls back to the range-checked path)

Hardening (no reproducer feasible; UB or fragile ordering):
- ALP/ALPRD decompression: uint8_t stack buffers were filled/read through
  uint16_t/uint32_t/uint64_t pointers (alignment + strict-aliasing UB) -
  use properly typed buffers
- ALPRD compression: same for the encoded-parts member arrays
- join hash table: the entry-publishing CAS used memory_order_acquire on
  success (must be release on ARM's weak memory model; x86's lock cmpxchg
  hides it) - release/acquire now, matching the sibling CAS
- WaitAll spin (COPY TO FILE) and GlobalUserSettings version gate:
  relaxed -> acquire loads
- Checksum(): raw uint64_t loads through reinterpret_cast -> memcpy-based Load
- UUIDv4 generation: uint32_t stores into std::array<uint8_t,16> ->
  memcpy-based Store
- StringUtil::CIHash: sign-extension of bytes >= 0x80 differed between
  signed-char (x86) and unsigned-char (ARM) platforms - zero-extend so the
  hash is platform-independent
- ctype calls on plain char (TryCastStringBool, ParseLogicalType, extension
  prefix, URI scheme detection) - cast to unsigned char (UB on negative chars)
- NumericHelper::SignedLength: signed overflow on INT64_MIN - compute in
  unsigned like the sibling FormatSigned
- BitwiseShiftLeft: guard 0 << 64 (shift by type width is UB)
- TryCastWithOverflowCheck: converting infinite floating-point bounds to
  integer types is UB - skip the bound comparison for non-integral DST
- SLURM memory limit parsing: guard double->idx_t overflow

Tests: test_fill_bounds, test_lead_lag_bounds, test_quantile_cont_bounds,
test_equi_width_bins_bounds (sqllogictest) + CIHash case (C++ unit test).
Full unittest suite passes on aarch64 (4949 test cases, 978147 assertions);
targeted repros verified clean under ASan+UBSan.

+++ src/common/string_util.cpp
@@ -445,7 +445,8 @@ uint64_t StringUtil::CIHash(const string &str) {
 uint64_t StringUtil::CIHash(const char *str, idx_t size) {
 	uint32_t hash = 0;
 	for (idx_t i = 0; i < size; i++) {
-		hash += static_cast<uint32_t>(StringUtil::CharacterToLower(static_cast<char>(str[i])));
+		// convert through uint8_t so the hash is identical on platforms with signed and unsigned char
+		hash += static_cast<uint32_t>(static_cast<uint8_t>(StringUtil::CharacterToLower(static_cast<char>(str[i]))));
 		hash += hash << 10;
 		hash ^= hash >> 6;
 	}
